requirements.txt
opencv-python
numpy
pillow
google-generativeai
pyaudio
SpeechRecognition
python-dotenv
streamlit
pyautogui